// web/lib/api/generated/model/calcMethodEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `PERCENT` - Percentage of taxable value
 * * `SPECIFIC_PER_UNIT` - Fixed amount per unit (reserved)
 */
export type CalcMethodEnum = typeof CalcMethodEnum[keyof typeof CalcMethodEnum];


export const CalcMethodEnum = {
  PERCENT: 'PERCENT',
  SPECIFIC_PER_UNIT: 'SPECIFIC_PER_UNIT',
} as const;