// web/lib/api/generated/model/cessType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { CalcMethodEnum } from './calcMethodEnum';

export interface CessType {
  readonly id: string;
  /** @maxLength 40 */
  code: string;
  /** @maxLength 120 */
  name: string;
  calc_method?: CalcMethodEnum;
  is_active?: boolean;
}